Abilites no longer instantiate, reset instead. Rain now has target marker and makes anyone wet within certain radius.

+++ Assets/_Divinity 2 VR/Scripts/ITCPlayerInputs.cs
@@ -13,7 +13,7 @@ namespace intheclouds
     public class ITCPlayerInputs : MonoBehaviour
     {
         public UserMenu menu;
-        public int holdTimeRequired = 1;
+        public float holdTimeRequired = 1;
         private float holdTimeLeftPrimaryButton;
         private bool primaryButtonTriggered;
         private GameManager gameManager;
